Try to get home page to load with Nav Bar
Alice down the rabbit hole. Getting a strange error in that the files are
not getting seen. No idea why. This is a bigger commit than expected but
it's all a learning exercise and will be written up. V. v. annoyed at this
right now so pushing and walking away.